split run pipeline dialog and form
