+++ apps/gitness/src/pages/pipeline-edit/context/data-store/reducer.ts
@@ -1,6 +1,8 @@
 import { DataActionName, DataActions, DataReducerState } from './types'
 
 export const initialState: DataReducerState = {
+  isInitialized: false,
+  //
   isDirty: false,
   isExistingPipeline: false,
   //
